+++ auth.js
@@ -3,229 +3,9 @@ const { notDeepStrictEqual } = require('assert');
 const { cp } = require('fs');
 const { maxHeaderSize } = require('http');
 const { send } = require('process');
-const TuyaApi = require('tuyapi');
 
 module.exports = function(RED) {
     "use strict";
-    function TuyaLocal(config) {
-        RED.nodes.createNode(this, config);
-    
-        const node = this;
-        node.config = config;
-        const tuyaDevice = new TuyaApi({
-            id: config.devId,
-            key: config.devKey,
-            ip: config.devIp,
-            version: config.protocolVer
-        });
-    
-        let tryReconnect = true;
-        let connectInterval = null;
-        let statusInterval = null;
-        let deviceInfo = { ip: config.devIp, name: config.devName, id: config.devId };
-       
-        function make_msg(event){
-            var available;
-            if (node.tuyaDevice === undefined){
-                available = false;
-            }else{
-                available = node.tuyaDevice.isConnected()
-            }
-            var msg = {}
-            if ("in_msg" in node){
-                msg = node.in_msg;
-                delete node.in_msg;
-            }else{
-                msg.called = false;
-             }
-        //node.warn(data);
-             if (!available){
-                 msg.error = event;
-             }
-             msg.data = {id:node.config.id, available: available, event: event };
-             return msg;
-        }
-
-        function send_msg(event){
-         
-            var msg = make_msg(event);
-            node.send(msg);
-        }
-        function connect(delay) {
-            //node.log(`Connecting to ${deviceInfo.name} @ ${deviceInfo.ip} (delay: ${delay ? 'yes' : 'no'})`)
-            clearTimeout(connectInterval);
-            clearTimeout(statusInterval);
-            if (tuyaDevice.isConnected()) {
-  
-                
-                return;
-            }
-            node.status({fill:"red",shape:"dot",text:"finding"});
-  
-            tuyaDevice.find({'options': {'timeout':3000}}).then( () => {
-                node.status({fill:"yellow",shape:"dot",text:"found"});
-                if (delay) {
-                    connectInterval = setTimeout(() => connect(), 5000);
-                    send_msg("connect on 5");
-                } else {
-                    if (tuyaDevice.isConnected()) {
-                        send_msg("already connected");
-                        return;
-                    }
-                    node.status({ fill: 'yellow', shape: 'dot', text: 'connecting...' });
-                    tuyaDevice.connect().then(() => { 
-                        send_msg("connec then");
-                    }).catch(() => { 
-                      send_msg("catch failed connect")
-                    });
-                }
-            }, (reason) => { 
-                node.status({fill:"red",shape:"ring",text:"find failed: " + reason+ "id:"+tuyaDevice.id+ "ip:"+tuyaDevice.ip});
-                send_msg("find failed")
-         
-            }); 
-           
-        }
-    
-        function disconnect() {
-            clearTimeout(connectInterval);
-            tryReconnect = false;
-            node.log(`Disconnect request for ${deviceInfo.name}`);
-            if (tuyaDevice.isConnected()) {
-                node.log(`Device connected, disconnecting...`);
-                tuyaDevice.disconnect();
-                node.log(`Disconnected`);
-            }
-            send_msg("disconnect");
-        
-        }
-    
-        function handleDisconnection() {
-            node.status({ fill: 'red', shape: 'ring', text: 'disconnected' });
- 
-            clearTimeout(statusInterval);
-            //node.log(`Device ${deviceInfo.name} disconnected, reconnect: ${tryReconnect}`);
-            if (tryReconnect) {
-               
-                connect(true);
-                 // go on to emit a message in case anyones listening
-            }
-            send_msg("disconnection");
-
-        }
-    
-        tuyaDevice.on('connected', () => {
-            //node.log(`Device ${deviceInfo.name} connected!`);
-            clearTimeout(connectInterval);
-            //if (config.pollingInterval !== 0) {
-             //   statusInterval = setInterval(() => {
-            //        tuyaDevice.get({ schema: true }).then(() => {}).catch(ex => {
-            //            node.log(`Error while polling status for ${deviceInfo.name}: ${ex.message}`);
-            //        });
-            //    }, config.pollingInterval * 1000);
-            //}
-            node.status({ fill: 'green', shape: 'dot', text: tuyaDevice.device.ip +  ` connected @ ${new Date().toLocaleTimeString()}` });
-            send_msg("connected")
-          
-          
-        });
-    
-        tuyaDevice.on('disconnected', () => {
-            //node.log(`Device ${deviceInfo.name} disconnected, reconnect: ${tryReconnect}`);
-            handleDisconnection();
-        });
-        tuyaDevice.on('error', (err) => {
-            //node.log(`Device ${deviceInfo.name} in error state: ${err}, reconnect: ${tryReconnect}`);
-            handleDisconnection();
-        });
-    
-        //7=requested response, 8=proactive update from device)
-        tuyaDevice.on('data', (data, commandByte) => {
-            if (commandByte == 8) {
-                //node.warn("proactive");
-                return;
-            }
-            var msg = make_msg("data")
-            msg.commandByte = commandByte;
-            msg.payload = data;
-          
-            node.send( msg);
-            
-        });
-    
-        node.on('input', (msg) => {
-            msg.called = true;
-            node.in_msg = msg;
-            let command = msg.payload;
-            if (typeof command === 'string') {
-                switch (command) {
-                    case 'request':
-                        if (tuyaDevice.isConnected()){
-                             try{
-                                tuyaDevice.get({ schema: true });
-                             }catch(e){
-                                send_msg(e)
-                            }
-                            return;
-                        }else{
-                            send_msg(`request not made as not connected ${deviceInfo.name}`);
-                            
-                        } 
-                        break;
-                    case 'request1':
-                        if (tuyaDevice.isConnected()){
-                            try{
-                               tuyaDevice.get({ schema: false });
-                               send_msg("got");
-                            }catch(e){
-                               send_msg(e)
-                           }
-                           return;
-                       }else{
-                          send_msg( `request1 not made as not connected ${deviceInfo.name}`);
-                         
-                       } 
-                        break;    
-                    case 'connect':
-                        //node.warn("connect command");
-                        connect();
-                        break;
-                    case 'disconnect':
-                        disconnect();
-                        break;
-                    case 'toggle':
-                        tuyaDevice.toggle();
-                        break;
-                }
-            } else if ( typeof command == "boolean" ) {
-				tuyaDevice.set({set: req}).then( () => {
-					node.status({fill:"green",shape:"dot",text: 'set success at:' + getHumanTimeStamp()});
-				}, (reason) => {
-					node.status({fill:"red",shape:"dot",text: 'set state failed:' + reason});
-				});
-
-			} else if ( "multiple" in command) {
-				tuyaDevice.set({
-					multiple:true,
-					data: command.data
-				});
-			} else if ('dps' in command) {
-                //node.warn("set "+command.dps);
-                tuyaDevice.set(command);
-            } else {
-                node.log(`Unknown command for ${deviceInfo.name}: ${command}`);
-            }
-        });
-    
-        node.on('close', (removed, done) => {
-            disconnect();
-            done();
-        });
-    
-        connect();
-    }
-	RED.nodes.registerType("tuya-local-socket",TuyaLocal);
-
     function tuya_auth(config) {
         RED.nodes.createNode(this, config);
 
@@ -489,652 +269,10 @@ module.exports = function(RED) {
     RED.nodes.registerType("tuya_sign", tuya_sign);
 
 
-    function TuyaDLocal(config) {
-        var node = this;
-        node.config = config;
-        RED.nodes.createNode(this, config);
-    
-        let tryReconnect = true;
-        let connectInterval = null;
-        let statusInterval = null;
-
-        function make_msg(event){
-            var available;
-            if (node.tuyaDevice === undefined){
-                available = false;
-            }else{
-                available = node.tuyaDevice.isConnected()
-            }
-            var msg = {}
-            if ("in_msg" in node){
-                msg = node.in_msg;
-                delete node.in_msg;
-            }else{
-                msg.called = false;
-             }
-        //node.warn(data);
-             if (!available){
-                 msg.error = event;
-             }
-             msg.data = {id:node.config.id, available: available, event: event };
-             return msg;
-        }
-
-        function send_msg(event){
-            node.status({fill:"yellow",shape:"dot",text:"output"});
-            var msg = make_msg(event);
-            node.send(msg);
-        }
-
-
-    
-        node.on('close', (removed, done) => {
-            disconnectOnClose();
-            done();
-        });
-    
-        node.on('input', (msg) => {
-            node.status({fill:"yellow",shape:"dot",text:"input: "+msg.payload});
-            msg.oldpayload=msg.payload;
-             function disconnectOnClose() {
-                clearTimeout(connectInterval);
-                tryReconnect = false;
-                node.log(`Disconnect on close request for ${node.config.name}`);
-                if (node.tuyaDevice.isConnected()) {
-                    node.log(`Device connected, disconnecting...`);
-                    node.tuyaDevice.disconnect();
-                }
-            }
-
-            function disconnect() {
-                clearTimeout(connectInterval);
-                tryReconnect = false;
-                node.log(`Disconnect request for ${node.config.name}`);
-                if (node.tuyaDevice.isConnected()) {
-                    node.log(`Device connected, disconnecting...`);
-                    node.tuyaDevice.disconnect();
-                    node.log(`Disconnected`);
-                }
-                send_msg("disconnected");
-                
-            }
-            config = node.config;
-            msg.called = true; // for the dynamic socket we are always called so make sure there is return node info
- 
-            node.in_msg = msg;
-
-            let command = msg.payload;
-            if (typeof command === 'string') {
-                switch (command) {
-                    case 'request':
-                        if (node.tuyaDevice!== undefined && node.tuyaDevice.isConnected()){
-                          try{
-                              node.status({fill:"yellow",shape:"dot",text:"get request"});
-                              node.tuyaDevice.get({ schema: true });
-                          }catch(e){
-                            
-                             send_msg(e);
-                            
-                            }
-                        }else{
-                          
-                            send_msg(`request not made as device disconnected ${node.config.name}` )
-                          
-                        }
-                        break;
-                    case 'request1':
-                        if (node.tuyaDevice!== undefined && node.tuyaDevice.isConnected()){
-                            try{
-                                node.status({fill:"yellow",shape:"dot",text:"get request"});
-                                node.tuyaDevice.get({ schema: false });
-                            }catch(e){
-                                node.status({fill:"red",shape:"dot",text:"get request failed"});
-                               send_msg(e)
-                               
-                              }
-                          }else{
-                              send_msg(`request not made as not connected ${node.config.name}`);
-                             
-                          }
-                          break;  
-                    case 'connect':{
-                        var config = node.config;
-                        //node.warn("id:"+ msg.id)
-                        if (msg.hasOwnProperty("ip")){
-                        
-                            config.devIp = msg.ip;
-                        }
-                        if (msg.hasOwnProperty("key")){
-                            config.devKey = msg.key;
-                        }
-                        if (msg.hasOwnProperty("id")){
-
-                            config.devId = msg.id;
-                        }
-                        const tuyaDevice = new TuyaApi({
-                            id: config.devId,
-                            key: config.devKey,
-                            ip: config.devIp,
-                            version: config.protocolVer
-                        });
-                        node.tuyaDevice = tuyaDevice;
-                        node.tuyaDevice.connecting = false;
-                        tuyaDevice.on('connected', () => {
-                            //node.log(`Device ${deviceInfo.name} connected!`);
-                            node.tuyaDevice.connecting=false;
-                            clearTimeout(connectInterval);
-                            //if (config.pollingInterval !== 0) {
-                             //   statusInterval = setInterval(() => {
-                            //        tuyaDevice.get({ schema: true }).then(() => {}).catch(ex => {
-                            //            node.log(`Error while polling status for ${deviceInfo.name}: ${ex.message}`);
-                            //        });
-                            //    }, config.pollingInterval * 1000);
-                            //}
-                            node.status({ fill: 'green', shape: 'dot', text: tuyaDevice.device.ip +  ` connected @ ${new Date().toLocaleTimeString()}` });
-                            send_msg("connected")
-    
-                        });
-                    
-                        tuyaDevice.on('disconnected', () => {
-                            //node.log(`Device ${deviceInfo.name} disconnected, reconnect: ${tryReconnect}`);
-                            handleDisconnection();
-                        });
-                        tuyaDevice.on('error', (err) => {
-                            //node.log(`Device ${deviceInfo.name} in error state: ${err}, reconnect: ${tryReconnect}`);
-                            handleDisconnection();
-                        });
-                        tuyaDevice.on('data', (data, commandByte) => {
-                            node.status({fill:"yellow",shape:"dot",text:"data received"});
-                            if (commandByte == 8) {
-                                //node.warn("proactive");
-                                return;
-                            }
-                           var msg = make_msg("data");
-                            msg.commandByte = commandByte;
-                            msg.payload = data;
-                            node.send( msg);
-                               
-                        }); 
-                        function connect(delay) {
-                            //node.log(`Connecting to ${deviceInfo.name} @ ${deviceInfo.ip} (delay: ${delay ? 'yes' : 'no'})`)
-                            clearTimeout(connectInterval);
-                            clearTimeout(statusInterval);
-                            
-                            if (tuyaDevice.isConnected()) {
-                                
-                                return;
-                            }else if (tuyaDevice.connecting){
-                                
-                                return  
-                            }
-                            tuyaDevice.connecting = true;
-                            node.status({fill:"red",shape:"dot",text:"finding"});
-                            tuyaDevice.find({'options': {'timeout':4000}}).then( () => {
-                                node.status({fill:"yellow",shape:"dot",text:"found"});
-                                if (delay) {
-                                    connectInterval = setTimeout(() => connect(), 5000);
-                                    send_msg("connect in 5");
-                                } else {
-                                    if (tuyaDevice.isConnected()) {
-                                        tuyaDevice.connecting = false;
-                                        send_msg("found already connected")
-                                        return;
-                                    }
-                                    node.status({ fill: 'yellow', shape: 'dot', text: 'connecting...' });
-                                    tuyaDevice.connect().then(() => { 
-                                        tuyaDevice.connecting = false;
-                                      
-                                       
-
-                                    }).catch(() => { 
-                                        tuyaDevice.connecting = false;
-                                        send_msg("connect catch")
-  
-                                    });
-                                }
-                            }, (reason) => { 
-                                tuyaDevice.connecting = false;
-                                node.status({fill:"red",shape:"ring",text:"find failed: " + reason + "id:"+tuyaDevice.device.id+ "ip:"+tuyaDevice.device.ip});
-                                send_msg("connect find")
-                            }); 
-                           
-                        }
-                    
- 
-                    
-                        function handleDisconnection() {
-                            clearTimeout(statusInterval);
-                            //node.log(`Device ${deviceInfo.name} disconnected, reconnect: ${tryReconnect}`);
-                            if (tryReconnect) {
-                                //node.warn("reconnect following disconnect");
-                                connect();
-                                // go on to emit message
-                            }
-                            node.status({ fill: 'red', shape: 'ring', text: 'disconnected: '+node.tuyaDevice.device.id });
-                            send_msg("disconnected");
-                          
-                        } 
-                        //node.warn("connect command");
-                        connect();
-                        break;
-                    }
-                    case 'disconnect':
-                        disconnect();
-                        break;
-                    case 'toggle':
-                        node.tuyaDevice.toggle();
-                        break;
-                    
-                }
-            } else if ( typeof command == "boolean" ) {
-				tuyaDevice.set({set: req}).then( () => {
-					node.status({fill:"green",shape:"dot",text: 'set success at:' + getHumanTimeStamp()});
-				}, (reason) => {
-					node.status({fill:"red",shape:"dot",text: 'set state failed:' + reason});
-				});
-
-			} else if ( "multiple" in command) {
-				tuyaDevice.set({
-					multiple:true,
-					data: command.data
-				});
-			} else if ('dps' in command) {
-                //node.warn("set "+command.dps);
-                tuyaDevice.set(command);
-            } else {
-                node.warn(`Unknown command for ${deviceInfo.name}: ${command}`);
-            }
-        });
-    
- 
-    
-    }
-	RED.nodes.registerType("tuya-local-dsocket",TuyaDLocal);
 
 
 
-    function TuyaMLocal(config) {
-        var node = this;
-        node.config = config;
-      
-      
-        RED.nodes.createNode(this, config);
-
-        node.tuyaDevices={};
-        node.in_msgs={};
-    
-        node.on('close', (removed, done) => {
-            Object.keys(node.tuyaDevices).forEach(function(key) {
-                disconnect(node.tuyaDevices[key]);
-               
-              });
-          
-            done();
-        });
-
-        function debug(tuyaDevice,message){
-            return;
-            var msg = {}
-    
-            if (tuyaDevice === undefined){
-              
-                return ;
-            }
-            var in_msg;
-            var id;
-            if (typeof tuyaDevice === 'string'){
-                id = tuyaDevice;
-
-            }else{
-               id = tuyaDevice.device.id;
-            }
-            in_msg = node.in_msgs[id];
-
-            if (in_msg !== undefined &&in_msg.hasOwnProperty("debug") && in_msg.debug == true && in_msg.hasOwnProperty("payload")){
-                node.warn(message+" "+in_msg.payload)
-            }else if (in_msg!== undefined){
-                if (in_msg.hasOwnProperty("debug") && in_msg.debug == true){
-                    node.warn(message)
-                }
-            }
-        }
-
-        function send_msg(tuyaDevice,event){
-         
-            debug(tuyaDevice,event);
-            var msg = make_msg(tuyaDevice,event);
-           
-            node.send(msg);
-            node.status({fill:"yellow",shape:"dot",text:"output "});
-        }
-        function make_msg(tuyaDevice,event){
-            var msg = {}
-
-            if (tuyaDevice === undefined){
-                msg.data = {id: "undefined",available: false, event:event};
-                return msg;
-            }
-            var in_msg;
-            var id;
-            if (typeof tuyaDevice === 'string'){
-                id = tuyaDevice;
-                in_msg = node.in_msgs[tuyaDevice]
-            }else{
-               id = tuyaDevice.device.id;
-            }
-            in_msg = node.in_msgs[id];
-
-            if (in_msg !== undefined){
-                msg = in_msg;
-                msg.called = true;
-                //node.warn("delete 788");
-                delete node.in_msgs[id];
-            }else{
-                //node.warn("no message in disconnect");
-                msg.called = false;
-            }
-            var available;
-            if  (typeof tuyaDevice === 'string'){
-                available = true;
-            }else{
-                available =  tuyaDevice.isConnected();
-            }
-            if (!available){
-                msg.error = event;
-            }
-            msg.data = { id: id,  available: available ,event: event };
-            return msg;
-        }
-    
-        node.on('input', (msg) => {
-           
-            node.status({fill:"yellow",shape:"dot",text:"input: "+msg.payload});
-            msg.oldpayload=msg.payload
-            if (msg.force === undefined && node.in_msgs[msg.id] !=undefined){
-                msg.error = "already processing message "+node.in_msgs[msg.id];
-                msg.old = node.in_msgs[msg.id];
-                if (msg.timestamp - msg.old.timestamp < 60000){
-                    node.status({fill:"red",shape:"dot",text:"eek "+msg.payload});
-                    //node.send(msg)
-                  
-                }
-            }
-
-   
-            function disconnect(tuyaDevice) {
-                debug(tuyaDevice,"disconnect");
-                clearTimeout(connectInterval);
-                tuyaDevice.tryReconnect = false;
-                node.log(`Disconnect request for ${tuyaDevice.device.id}`);
-                if (tuyaDevice.isConnected()) {
-                    log(`Device connected, disconnecting...`);
-                    tuyaDevice.disconnect();
-                }
-
-                send_msg(tuyaDevice,"disconnect");
-                
-            }
-
-            
-   
-            msg.called = true; // for the dynamic socket we are always called so make sure there is return node info
- 
-  
-
-            let command = msg.payload;
-
-            if (command != "status"){
-                if (msg.id === undefined){
-
-                    node.warn("missing device id");
-                    node.send(msg);
-                    return;
-                 }
-
-                if (msg.key === undefined){
-
-                    node.warn("missing device key");
-                    node.send(msg);
-                    return;
-                  }
-           }else{
-                msg.id="status";
-           }
-           var tuyaDevice = node.tuyaDevices[msg.id];
-
-
-            node.in_msgs[msg.id] = msg;
-
-
-            //node.warn("check command");
-            if (typeof command === 'string') {
-                switch (command) {
-                    case 'request':
-                        debug(tuyaDevice,"request");
-                        if (tuyaDevice!== undefined && tuyaDevice.isConnected()){
-                          try{
-                                node.status({fill:"yellow",shape:"dot",text:"get request"});
-                              tuyaDevice.get({ schema: true });
-                              //send_msg(tuyaDevice,"request got");
-                          }catch(e){
-
-                            node.status({fill:"red",shape:"dot",text:"get request failed: "+tuyaDevice.isConnected()});
-                             send_msg(tuyaDevice,"request error");
-                            
-                            }
-                        }else if (tuyaDevice === undefined){ 
-                            node.error =`missing device ${msg.id}`;
-                            node.warn(node.error);
-                            msg.data = {available: false, id: msg.id}
-                            node.send(msg);
-                        }else{
-                            send_msg(tuyaDevice, `request not made as not connected ${msg.id}`);
-                          
-                        }
-                        break;
-                    case 'request1':
-                        if (tuyaDevice!== undefined && tuyaDevice.isConnected()){
-                            try{
-                                node.status({fill:"yellow",shape:"dot",text:"get request1"});
-                                tuyaDevice.get({ schema: false });
-                                //send_msg(tuyaDevice,"request1 got");
-                            }catch(e){
-                                node.status({fill:"red",shape:"dot",text:"get request1 failed: "+tuyaDevice.isConnected()});
-                                send_msg(tuyaDevice,"request1 error");
-                               
-                              }
-                          }else if (tuyaDevice === undefined){ 
-                            node.error =`missing device ${msg.id}`;
-                            node.warn(node.error);
-                            msg.data = {available: false, id: msg.id}
-                            node.send(msg);
-                        }else{
-                            send_msg(tuyaDevice,`request1 not made as not connected ${msg.id}`);
-                             
-                          }
-                          break;  
-                    case 'status'  :
-                        var status={};  
-                        Object.keys(node.tuyaDevices).forEach(function(key) {
-                            var tuyaDevice = node.tuyaDevices[key];
-                            status[tuyaDevice.device.id]= {ip: tuyaDevice.device.ip,available: tuyaDevice.isConnected()}
-                         });
-                         node.warn("status collated")
-                         var msg = make_msg("status","status");
-                         node.warn("status message made")
-                         msg.payload=status;
-                         node.send(msg)
-                          break;                          
-                    case 'connect':{
-                        var config = node.config;
-        
-                        if (msg.reconnect === undefined && node.tuyaDevices[ msg.id] !== undefined){
-                            if (node.tuyaDevices[ msg.id].isConnected()){
-                                send_msg(node.tuyaDevices[ msg.id],"already connected");
-                                return;
-                            }else{
-                                node.tuyaDevices[ msg.id].tryReconnect = false;
-                                node.tuyaDevices[ msg.id].disconnect();
-                                delete node.tuyaDevices[ msg.id];
-                                //node.warn("delete old device");
-                               
-                            }
-                        }
-       
-                        const tuyaDevice = new TuyaApi({
-                            id: msg.id,
-                            key: msg.key,
-                            ip: msg.ip,
-                            version: config.protocolVer
-                        });
-                        tuyaDevice.tryReconnect = true;
-                        tuyaDevice.connectInterval =null;
-                        tuyaDevice.statusInterval =null;
-                        tuyaDevice.connecting = false;
-                        node.tuyaDevices[ msg.id] = tuyaDevice;
-                        tuyaDevice.on('connected', () => {
-                            //node.log(`Device ${deviceInfo.name} connected!`);
-                            clearTimeout(tuyaDevice.connectInterval);
-                            tuyaDevice.tryReconnect = false;
-                            tuyaDevice.connectInterval =null;
-                            tuyaDevice.statusInterval =null;
-                            tuyaDevice.connecting=false;
-                            //if (config.pollingInterval !== 0) {
-                             //   statusInterval = setInterval(() => {
-                            //        tuyaDevice.get({ schema: true }).then(() => {}).catch(ex => {
-                            //            node.log(`Error while polling status for ${deviceInfo.name}: ${ex.message}`);
-                            //        });
-                            //    }, config.pollingInterval * 1000);
-                            //}
-                            node.status({ fill: 'green', shape: 'dot', text: tuyaDevice.device.ip +  ` connected @ ${new Date().toLocaleTimeString()}` });
-                            send_msg(tuyaDevice,"connected");
-    
-                        });
-                    
-                        tuyaDevice.on('disconnected', () => {
-                            //node.log(`Device ${deviceInfo.name} disconnected, reconnect: ${tryReconnect}`);
-                            debug("seen disconnected")
-                            handleDisconnection(tuyaDevice);
-                        });
-                        tuyaDevice.on('error', (err) => {
-                            //node.log(`Device ${deviceInfo.name} in error state: ${err}, reconnect: ${tryReconnect}`);
-                            debug("seen error")
-                            handleDisconnection(tuyaDevice);
-                        });
-                        tuyaDevice.on('data', (data, commandByte) => {
-                            node.status({fill:"yellow",shape:"dot",text:"data received"});
-                            if (commandByte == 8) {
-                                node.warn("proactive");
-                                return;
-                            }
-                            var msg = make_msg(tuyaDevice,true,"data")
-                            msg.commandByte = commandByte;
-                            msg.payload = data;
-                            node.send( msg);
-
-                        }); 
-                        function connect(tuyaDevice,delay) {
-                            //node.log(`Connecting to ${deviceInfo.name} @ ${deviceInfo.ip} (delay: ${delay ? 'yes' : 'no'})`)
-                            debug(tuyaDevice,"connect")
-                            clearTimeout(tuyaDevice.connectInterval);
-                            clearTimeout(tuyaDevice.statusInterval);
-                            if (tuyaDevice.isConnected()) {
-                               
-                                return;
-                            }else if (tuyaDevice.connecting){
-                               
-                                return  
-                            }
-
-                            debug(tuyaDevice,"connect finding")
-                            node.status({fill:"red",shape:"dot",text:"finding"});
-                            tuyaDevice.connecting=true;
-                            tuyaDevice.find({'options': {'timeout':4000}}).then( () => {
-                                node.status({fill:"yellow",shape:"dot",text:"found"});
-                                if (delay) {
-                                    node.status({fill:"yellow",shape:"dot",text:"delayed reconnect"});
-                                    tuyaDevice.connectInterval = setTimeout(() => connect(tuyaDevice), 5000);
-                                    send_msg(tuyaDevice,"connect in 5");
-                                } else {
-                                    if (tuyaDevice.isConnected()) {
-                                        send_msg(tuyaDevice,"already connected");
-                                        return;
-                                    }
-                                    node.status({ fill: 'yellow', shape: 'dot', text: 'connecting...' });
-                                
-                                    tuyaDevice.connect().then(() => { 
-                                        //send_msg(tuyaDevice,"connect then");
-                                        tuyaDevice.connecting=false;
-                                       
-                                        //node.warn("connect then block")
-                                    }).catch((e) => { 
-                                        //node.warn("catch connect");
-                                        tuyaDevice.connecting=false;
-  
-                                        send_msg(tuyaDevice,"failed connect");    
-                                    });
-                                }
-                            }, (reason) => { 
-                                tuyaDevice.connecting=false;
-                                send_msg(tuyaDevice,"find failed");
-                                node.status({fill:"red",shape:"ring",text:"find failed: " + reason + "id:"+tuyaDevice.device.id+ "ip:"+tuyaDevice.device.ip});
-                            }); 
-                           
-                        }
-                    
- 
-                    
-                        function handleDisconnection(tuyaDevice) {
-                            debug(tuyaDevice,"disconnect")
-                            clearTimeout(tuyaDevice.statusInterval);
-                            send_msg(tuyaDevice,"disconnected");
-                            node.status({ fill: 'red', shape: 'ring', text: 'disconnected'+tuyaDevice.device.id });
-                            if (tuyaDevice.tryReconnect) {
-                                //node.warn("reconnect following disconnect");
-                                debug(tuyaDevice,"retry")
-                                connect(tuyaDevice,true);
-                            }else{
-                                debug(tuyaDevice,"no retry")
-                            }
-                            
-                           
-                        } 
-                        //
-                        //node.warn("call connect");
-                        connect(tuyaDevice,0);
-                        break;
-                    }
-                    case 'disconnect':
-                        disconnect(tuyaDevice);
-                        break;
-                    case 'toggle':
-                        tuyaDevice.toggle();
-                        break;
-                    
-                }
-            } else if ( typeof command == "boolean" ) {
-				tuyaDevice.set({set: req}).then( () => {
-					node.status({fill:"green",shape:"dot",text: 'set success at:' + getHumanTimeStamp()});
-				}, (reason) => {
-					node.status({fill:"red",shape:"dot",text: 'set state failed:' + reason});
-				});
-
-			} else if ( "multiple" in command) {
-				tuyaDevice.set({
-					multiple:true,
-					data: command.data
-				});
-			} else if ('dps' in command) {
-                //node.warn("set "+command.dps);
-                tuyaDevice.set(command);
-            } 
-        });
-    
-        //debug("drop through");
-    
-    }
-	RED.nodes.registerType("tuya-local-msocket",TuyaMLocal);
-};
+ };
 
 
 /**
